switchboard tokens support
