+++ solauto-sdk/local/updateMarginfiLUT.ts
@@ -15,7 +15,6 @@ async function addBanks() {
       const accounts = MARGINFI_ACCOUNTS[group][key];
       await updateLookupTable(
         [
-          key,
           accounts.bank,
           accounts.liquidityVault,
           accounts.vaultAuthority,
