+++ solauto-sdk/local/updateSolautoLUT.ts
@@ -15,6 +15,7 @@ import {
   getAllMarginfiAccountsByAuthority,
   getSolanaRpcConnection,
 } from "../src/utils";
+import { SWITCHBOARD_PRICE_FEED_IDS } from "../src/constants/switchboardConstants";
 
 const LOOKUP_TABLE_ADDRESS = new PublicKey(SOLAUTO_LUT);
 const solautoManagerTokenAccounts = getTokenAccounts(
@@ -30,23 +31,22 @@ export async function updateSolautoLut(additionalAccounts?: string[]) {
   const [connection, umi] = getSolanaRpcConnection(
     buildHeliusApiUrl(process.env.HELIUS_API_KEY!)
   );
-  const solautoManagerMarginfiAccounts =
-    await getAllMarginfiAccountsByAuthority(
-      connection,
-      umi,
-      SOLAUTO_MANAGER,
-      false
-    );
+  const ismAccounts = await getAllMarginfiAccountsByAuthority(
+    connection,
+    umi,
+    SOLAUTO_MANAGER,
+    false
+  );
 
   return updateLookupTable(
     [
       ...STANDARD_LUT_ACCOUNTS,
+      ...ALL_SUPPORTED_TOKENS,
       ...solautoManagerTokenAccounts.map((x) => x.toString()),
       ...solautoFeeWalletTokenAccounts.map((x) => x.toString()),
+      ...ismAccounts.map((x) => x.marginfiAccount.toString()),
+      ...Object.values(SWITCHBOARD_PRICE_FEED_IDS),
       ...(additionalAccounts ?? []),
-      ...solautoManagerMarginfiAccounts.map((x) =>
-        x.marginfiAccount.toString()
-      ),
     ],
     LOOKUP_TABLE_ADDRESS
   );
