+++ solauto-sdk/src/constants/README.md
@@ -1,5 +1,3 @@
-TODO: move this to main readme in root folder when that's ready
-
 Instructions for supporting new token:
 
 - add public key to tokenConstants.ts
