+++ solauto-sdk/src/utils/generalUtils.ts
@@ -1,24 +1,5 @@
-import {
-  Connection,
-  PublicKey,
-  Transaction,
-  VersionedTransaction,
-} from "@solana/web3.js";
-import {
-  MaybeRpcAccount,
-  publicKey,
-  Umi,
-  PublicKey as UmiPublicKey,
-} from "@metaplex-foundation/umi";
-import { PYTH_PRICE_FEED_IDS } from "../constants/pythConstants";
-import { fromBaseUnit, toBaseUnit } from "./numberUtils";
-import { PRICES } from "../constants/solautoConstants";
-import {
-  PullFeed,
-} from "@switchboard-xyz/on-demand";
-import { AnchorProvider, Idl, Program } from "@coral-xyz/anchor";
-import switchboardIdl from "../idls/switchboard.json";
-import { SWITCHBOARD_PRICE_FEED_IDS } from "../constants/switchboardConstants";
+import { PublicKey } from "@solana/web3.js";
+import { MaybeRpcAccount, publicKey, Umi } from "@metaplex-foundation/umi";
 
 export function consoleLog(...args: any[]): void {
   if ((globalThis as any).LOCAL_TEST) {
@@ -77,137 +58,6 @@ export function zip<T, U>(list1: T[], list2: U[]): [T, U][] {
   }
 
   return result;
-}
-
-export async function fetchTokenPrices(
-  conn: Connection,
-  mints: PublicKey[]
-): Promise<number[]> {
-  const currentTime = currentUnixSeconds();
-  if (
-    !mints.some(
-      (mint) =>
-        !(mint.toString() in PRICES) ||
-        currentTime - PRICES[mint.toString()].time > 3
-    )
-  ) {
-    return mints.map((mint) => PRICES[mint.toString()].price);
-  }
-
-  const pythMints = mints.filter((x) => x.toString() in PYTH_PRICE_FEED_IDS);
-  const switchboardMints = mints.filter(
-    (x) => x.toString() in SWITCHBOARD_PRICE_FEED_IDS
-  );
-
-  const [pythData, switchboardData] = await Promise.all([
-    zip(pythMints, await getPythPrices(pythMints)),
-    zip(switchboardMints, await getSwitchboardPrices(conn, switchboardMints)),
-  ]);
-
-  const prices = mints.map((mint) => {
-    const item = [...pythData, ...switchboardData].find((data) =>
-      data[0].equals(mint)
-    );
-    return item ? item[1] : 0;
-  });
-
-  for (var i = 0; i < mints.length; i++) {
-    PRICES[mints[i].toString()] = {
-      price: prices[i],
-      time: currentUnixSeconds(),
-    };
-  }
-
-  return prices;
-}
-
-export async function getPythPrices(mints: PublicKey[]) {
-  if (mints.length === 0) {
-    return [];
-  }
-
-  const priceFeedIds = mints.map(
-    (mint) => PYTH_PRICE_FEED_IDS[mint.toString()]
-  );
-
-  const getReq = async () =>
-    await fetch(
-      `https://hermes.pyth.network/v2/updates/price/latest?${priceFeedIds.map((x) => `ids%5B%5D=${x}`).join("&")}`
-    );
-
-  const prices: number[] = await retryWithExponentialBackoff(
-    async () => {
-      let resp = await getReq();
-      let status = resp.status;
-      if (status !== 200) {
-        throw new Error(JSON.stringify(resp));
-      }
-
-      const json = await resp.json();
-      const prices = json.parsed.map((x: any) => {
-        if (x.price.expo > 0) {
-          return Number(toBaseUnit(Number(x.price.price), x.price.expo));
-        } else if (x.price.expo < 0) {
-          return fromBaseUnit(BigInt(x.price.price), Math.abs(x.price.expo));
-        } else {
-          return Number(x.price.price);
-        }
-      });
-
-      return prices;
-    },
-    5,
-    200
-  );
-
-  return prices;
-}
-
-export async function getSwitchboardPrices(
-  conn: Connection,
-  mints: PublicKey[]
-) {
-  if (mints.length === 0) {
-    return [];
-  }
-
-  const dummyWallet = {
-    publicKey: new PublicKey("11111111111111111111111111111111"),
-    signTransaction: async <T extends Transaction | VersionedTransaction>(
-      tx: T
-    ): Promise<T> => tx,
-    signAllTransactions: async <T extends Transaction | VersionedTransaction>(
-      txs: T[]
-    ): Promise<T[]> => txs,
-  };
-  const provider = new AnchorProvider(
-    conn,
-    dummyWallet,
-    AnchorProvider.defaultOptions()
-  );
-  const program = new Program(switchboardIdl as Idl, provider);
-
-  const results = await Promise.all(
-    mints.map(async (mint) => {
-      const feed = new PullFeed(
-        program,
-        new PublicKey(SWITCHBOARD_PRICE_FEED_IDS[mint.toString()])
-      );
-      const result = await feed.loadData();
-      return Number(result.result.value) / Math.pow(10, 18);
-    })
-  );
-
-  return results;
-}
-
-export function safeGetPrice(
-  mint: PublicKey | UmiPublicKey | undefined
-): number | undefined {
-  if (mint && mint?.toString() in PRICES) {
-    return PRICES[mint!.toString()].price;
-  }
-  return undefined;
 }
 
 export type ErrorsToThrow = Array<new (...args: any[]) => Error>;
