+++ solauto-sdk/src/utils/marginfiUtils.ts
@@ -9,11 +9,7 @@ import {
   safeFetchBank,
   safeFetchMarginfiAccount,
 } from "../marginfi-sdk";
-import {
-  currentUnixSeconds,
-  fetchTokenPrices,
-  safeGetPrice,
-} from "./generalUtils";
+import { currentUnixSeconds } from "./generalUtils";
 import {
   bytesToI80F48,
   fromBaseUnit,
@@ -30,6 +26,7 @@ import { PositionState, PositionTokenUsage } from "../generated";
 import { USD_DECIMALS } from "../constants/generalAccounts";
 import { LivePositionUpdates } from "./solauto/generalUtils";
 import { TOKEN_INFO } from "../constants";
+import { fetchTokenPrices, safeGetPrice } from "./priceUtils";
 
 interface AllMarginfiAssetAccounts extends MarginfiAssetAccounts {
   mint: PublicKey;
@@ -87,7 +84,6 @@ export function calcMaxLtvAndLiqThreshold(
 }
 
 export async function getMaxLtvAndLiqThreshold(
-  conn: Connection,
   umi: Umi,
   marginfiGroup: PublicKey,
   supply: {
@@ -128,7 +124,7 @@ export async function getMaxLtvAndLiqThreshold(
   }
 
   if (!supplyPrice) {
-    const [price] = await fetchTokenPrices(conn, [
+    const [price] = await fetchTokenPrices([
       toWeb3JsPublicKey(supply.bank!.mint),
     ]);
     supplyPrice = price;
@@ -175,7 +171,7 @@ export async function getAllMarginfiAccountsByAuthority(
     const positionStates = await Promise.all(
       marginfiAccounts.map(async (x) => ({
         publicKey: x.publicKey,
-        state: await getMarginfiAccountPositionState(conn, umi, {
+        state: await getMarginfiAccountPositionState(umi, {
           pk: toWeb3JsPublicKey(x.publicKey),
         }),
       }))
@@ -206,7 +202,6 @@ export async function getAllMarginfiAccountsByAuthority(
 }
 
 async function getTokenUsage(
-  conn: Connection,
   bank: Bank | null,
   isAsset: boolean,
   shares: number,
@@ -217,7 +212,7 @@ async function getTokenUsage(
   let marketPrice = 0;
 
   if (bank !== null) {
-    [marketPrice] = await fetchTokenPrices(conn, [
+    [marketPrice] = await fetchTokenPrices([
       toWeb3JsPublicKey(bank.mint),
     ]);
     const [assetShareValue, liabilityShareValue] =
@@ -275,7 +270,6 @@ interface BankSelection {
 type BanksCache = { [group: string]: { [mint: string]: Bank } };
 
 export async function getMarginfiAccountPositionState(
-  conn: Connection,
   umi: Umi,
   protocolAccount: { pk: PublicKey; data?: MarginfiAccount },
   marginfiGroup?: PublicKey,
@@ -360,7 +354,6 @@ export async function getMarginfiAccountPositionState(
         supply.mint = toWeb3JsPublicKey(supplyBank!.mint);
       }
       supplyUsage = await getTokenUsage(
-        conn,
         supplyBank!,
         true,
         bytesToI80F48(supplyBalances[0].assetShares.value),
@@ -378,7 +371,6 @@ export async function getMarginfiAccountPositionState(
         debt.mint = toWeb3JsPublicKey(debtBank!.mint);
       }
       debtUsage = await getTokenUsage(
-        conn,
         debtBank!,
         false,
         bytesToI80F48(debtBalances[0].liabilityShares.value),
@@ -393,7 +385,6 @@ export async function getMarginfiAccountPositionState(
 
   if (!supplyUsage) {
     supplyUsage = await getTokenUsage(
-      conn,
       supplyBank,
       true,
       0,
@@ -410,7 +401,6 @@ export async function getMarginfiAccountPositionState(
 
   if (!debtUsage) {
     debtUsage = await getTokenUsage(
-      conn,
       debtBank,
       false,
       0,
@@ -420,7 +410,6 @@ export async function getMarginfiAccountPositionState(
 
   const supplyPrice = safeGetPrice(supply.mint!)!;
   let [maxLtv, liqThreshold] = await getMaxLtvAndLiqThreshold(
-    conn,
     umi,
     marginfiGroup ?? new PublicKey(DEFAULT_MARGINFI_GROUP),
     {
